+++ src/types/components/event.ts
@@ -3,7 +3,7 @@ import Alarm from './alarm'
 import { RRule } from 'rrule'
 
 type Event = {
-    dtStamp: string | ComplexDate
+    dtStamp?: string | ComplexDate
     uid: string
     start?: string | ComplexDate
     class?: Classification