+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [1.0.0] - 2020-09-10
+
+- Added citation for regex from Ruby iCalendar library
+- Added rfc link to readme
+- Updated dtStamp as optional, current UTC time used when not set
+- Added schedule status event parameter, dquote parsing
+- Updated delegatedTo/From parsing to remove 'mailto:'
+- Updated datetime to expect iCalendar formatted strings
+
 ## [0.4.0] - 2020-05-07
 
 ### Added